% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerillo_excel.R
\name{cerillo_excel}
\alias{cerillo_excel}
\title{cerillo_excel}
\usage{
cerillo_excel(filename = paste0(Sys.Date(), ".xlsx"))
}
\arguments{
\item{filename}{The excel file that contains your OD600 readings in the 'dat' sheet and the variable data on the 'map' sheet}
}
\description{
cerillo_excel
}
