% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(project_name, location = getwd())
}
\arguments{
\item{project_name}{Generate a new project and subfolders}
}
\description{
create_project
}
