% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hpline.R
\docType{data}
\name{geom_hpline}
\alias{geom_hpline}
\alias{GeomHpline}
\alias{geom_vpline}
\alias{GeomVpline}
\title{Draw point-like short line segments}
\usage{
geom_hpline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_vpline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
The geoms `geom_hpline()` and `geom_vpline()` can be used as a drop-in
replacement for [`geom_point()`] but draw horizontal or vertical lines
(point-lines, or plines) instead of points. These lines can often be useful to
indicate specific parameter estimates in a plot. The geoms take position
aesthetics as `x` and `y` like [`geom_point()`], and they use `width` or `height`
to set the length of the line segment. All other aesthetics (`colour`, `size`,
`linetype`, etc.) are inherited from [`geom_segment()`].
}
\examples{
library(ggplot2)
ggplot(iris, aes(Species, Sepal.Length)) +
  geom_hpline(stat = "summary")

ggplot(iris, aes(Species, Sepal.Length)) +
  geom_point(position = "jitter", size = 0.5) +
  stat_summary(aes(colour = Species), geom = "hpline", width = 0.6, size = 1.5)

ggplot(iris, aes(Sepal.Length, Species, color = Species)) +
  geom_point(color = "grey50", alpha = 0.3, size = 2) +
  geom_vpline(data = sampler(5, 1, group = Species), height = 0.4) +
  scale_color_brewer(type = "qual", palette = 2, guide = "none") +
  facet_wrap(~.draw) +
  theme_bw()
}
\keyword{datasets}
