% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cerillo.R
\name{load_cerillo}
\alias{load_cerillo}
\title{load_cerillo}
\usage{
load_cerillo(filename, Time_int = 10, range = "F:CW")
}
\arguments{
\item{filename}{The excel file that contains your OD600 readings in the 'dat' sheet and the variable data on the 'map' sheet}

\item{Time_int}{Time between readings. Defaults to 10 mins}
}
\description{
load_cerillo
}
