% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LS}
\alias{LS}
\title{Convert line sizes measured as points to ggplot line sizes}
\usage{
LS(x)
}
\arguments{
\item{x}{numeric vector giving the line sizes in points}
}
\value{
numeric vector of ggplot line sizes
}
\description{
Converts line sizes measured as points (as given by Adobe Illustrator etc.) to ggplot line sizes
}
\keyword{internal}
