% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pp.ggsave}
\alias{pp.ggsave}
\title{default ggsave for PowerPoint figs}
\usage{
pp.ggsave(filename = default_name(plot), height = 10, width = 5.625, ...)
}
\description{
saves at sensible default for PowerPoint
}
\keyword{internal}
