% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.R
\name{scale_color_pal}
\alias{scale_color_pal}
\title{Palette scale helpers (avoid using global options)
Add these to a plot to apply your discrete palettes explicitly.}
\usage{
scale_color_pal(choice = c("pal1", "pal2", "pal3"))
}
\description{
Palette scale helpers (avoid using global options)
Add these to a plot to apply your discrete palettes explicitly.
}
