% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.R
\name{use_gold_theme}
\alias{use_gold_theme}
\title{Convenience: set theme globally and update geom defaults (opt-in)}
\usage{
use_gold_theme(
  colPal = c("pal1", "pal2", "pal3"),
  base_family = c("Arial", "Arial Narrow"),
  guides = FALSE
)
}
\arguments{
\item{colPal}{One of "pal1","pal2","pal3".}

\item{base_family}{"Arial" or "Arial Narrow".}

\item{guides}{Logical; TRUE adds subtle gridlines.}
}
\description{
Convenience: set theme globally and update geom defaults (opt-in)
}
